requirements.txt
pandas
numpy
matplotlib
scikit-learn
psycopg2-binary
sqlalchemy
